Create SingleNodeConnCB class for RPC tests
